Implement Sort interface for MentionRoute Router
- Move methods back into router file
- Remove commented out line
